cleanup and code style changes
